# swagger.json
{"openapi":"3.0.0","paths":{"/discord/command":{"post":{"operationId":"DiscordController_executeCommand","summary":"Ejecutar un comando del bot de Discord","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandDto"}}}},"responses":{"200":{"description":"Comando ejecutado exitosamente"},"400":{"description":"Datos inválidos"},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"},"500":{"description":"Error del servidor"}},"tags":["discord"],"security":[{"API-KEY":[]}]}},"/discord/health":{"get":{"operationId":"DiscordController_healthCheck","summary":"Verificar el estado del servicio","parameters":[],"responses":{"200":{"description":"Servicio funcionando correctamente"}},"tags":["discord"],"security":[{"API-KEY":[]}]}},"/birthdays":{"post":{"operationId":"BirthdayController_create","summary":"Crear un nuevo registro de cumpleaños","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBirthdayDto"}}}},"responses":{"201":{"description":"Cumpleaños creado exitosamente","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Birthday"}}}},"400":{"description":"Datos inválidos"},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"},"500":{"description":"Ese usuario ya existe"}},"tags":["birthdays"],"security":[{"API-KEY":[]}]},"get":{"operationId":"BirthdayController_findAll","summary":"Obtener todos los cumpleaños","parameters":[{"name":"guildId","required":true,"in":"query","description":"ID del servidor de Discord","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","description":"Número de página para paginación","schema":{"type":"number"}}],"responses":{"200":{"description":"Lista de cumpleaños","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Birthday"}}}}},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"}},"tags":["birthdays"],"security":[{"API-KEY":[]}]}},"/birthdays/upcoming":{"get":{"operationId":"BirthdayController_getUpcomingBirthdays","summary":"Obtener próximos cumpleaños de un servidor","parameters":[{"name":"guildId","required":true,"in":"query","description":"ID del servidor de Discord","schema":{"type":"string"}}],"responses":{"200":{"description":"Lista de próximos cumpleaños","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Birthday"}}}}},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"}},"tags":["birthdays"],"security":[{"API-KEY":[]}]}},"/birthdays/today":{"get":{"operationId":"BirthdayController_getTodaysBirthdays","summary":"Obtener cumpleaños del día actual de un servidor","parameters":[{"name":"guildId","required":true,"in":"query","description":"ID del servidor de Discord","schema":{"type":"string"}}],"responses":{"200":{"description":"Lista de cumpleaños de hoy","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Birthday"}}}}},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"}},"tags":["birthdays"],"security":[{"API-KEY":[]}]}},"/birthdays/{userId}":{"get":{"operationId":"BirthdayController_findOne","summary":"Obtener cumpleaños por ID de usuario","parameters":[{"name":"userId","required":true,"in":"path","description":"ID del usuario de Discord","schema":{"type":"string"}}],"responses":{"200":{"description":"Datos del cumpleaños","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Birthday"}}}},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"},"404":{"description":"Usuario no encontrado"}},"tags":["birthdays"],"security":[{"API-KEY":[]}]},"patch":{"operationId":"BirthdayController_update","summary":"Actualizar cumpleaños por ID de usuario","parameters":[{"name":"userId","required":true,"in":"path","description":"ID del usuario de Discord","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBirthdayDto"}}}},"responses":{"200":{"description":"Cumpleaños actualizado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Birthday"}}}},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"},"404":{"description":"Usuario no encontrado"}},"tags":["birthdays"],"security":[{"API-KEY":[]}]},"delete":{"operationId":"BirthdayController_remove","summary":"Eliminar cumpleaños por ID de usuario","parameters":[{"name":"userId","required":true,"in":"path","description":"ID del usuario de Discord","schema":{"type":"string"}}],"responses":{"200":{"description":"Cumpleaños eliminado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Birthday"}}}},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"},"404":{"description":"Usuario no encontrado"}},"tags":["birthdays"],"security":[{"API-KEY":[]}]}},"/openai/generate":{"post":{"operationId":"OpenaiController_generateResponse","summary":"Generar respuesta de OpenAI","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateRequestDto"}}}},"responses":{"200":{"description":"Respuesta generada exitosamente"},"400":{"description":"Datos inválidos"},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"},"500":{"description":"Error del servidor"}},"tags":["openai"],"security":[{"API-KEY":[]}]}},"/openai/config/{guildId}":{"get":{"operationId":"OpenaiController_getGuildConfig","summary":"Obtener configuración de IA de un servidor","parameters":[{"name":"guildId","required":true,"in":"path","description":"ID del servidor de Discord","schema":{"type":"string"}}],"responses":{"200":{"description":"Configuración obtenida exitosamente","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuildAIConfig"}}}},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"},"500":{"description":"Error del servidor"}},"tags":["openai"],"security":[{"API-KEY":[]}]},"patch":{"operationId":"OpenaiController_updateGuildConfig","summary":"Actualizar configuración de IA de un servidor","parameters":[{"name":"guildId","required":true,"in":"path","description":"ID del servidor de Discord","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAIConfigDto"}}}},"responses":{"200":{"description":"Configuración actualizada exitosamente","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuildAIConfig"}}}},"400":{"description":"Datos inválidos"},"401":{"description":"No autorizado - API Key inválida"},"403":{"description":"Prohibido - IP no autorizada"},"500":{"description":"Error del servidor"}},"tags":["openai"],"security":[{"API-KEY":[]}]}}},"info":{"title":"Discord Bot API","description":"API para el backend del bot de Discord con servicios de cumpleaños, OpenAI y más","version":"1.0","contact":{}},"tags":[{"name":"discord","description":""},{"name":"birthdays","description":""},{"name":"openai","description":""}],"servers":[],"components":{"securitySchemes":{"API-KEY":{"scheme":"bearer","bearerFormat":"API Key","type":"http","name":"Authorization","description":"Introduce tu API Key","in":"header"}},"schemas":{"CommandDto":{"type":"object","properties":{"command":{"type":"string","description":"Comando a ejecutar","example":"gpt","enum":["birthday","todaysBirthdays","gpt","aiConfig"]},"params":{"type":"object","description":"Parámetros del comando","example":{"prompt":"¿Cuál es la capital de Francia?","guildId":"987654321098765432","userId":"123456789012345678","channelName":"general"}}},"required":["command","params"]},"CreateBirthdayDto":{"type":"object","properties":{"userId":{"type":"string","description":"ID del usuario de Discord","example":"123456789012345678"},"username":{"type":"string","description":"Nombre de usuario de Discord","example":"JohnDoe#1234"},"guildId":{"type":"string","description":"ID del servidor de Discord","example":"987654321098765432"},"birthDate":{"format":"date-time","type":"string","description":"Fecha de nacimiento","example":"2000-01-01T00:00:00.000Z"},"notificationEnabled":{"type":"boolean","description":"Indica si las notificaciones están habilitadas","example":true,"default":false}},"required":["userId","username","guildId","birthDate"]},"Birthday":{"type":"object","properties":{"userId":{"type":"string","description":"ID del usuario de Discord","example":"123456789012345678"},"username":{"type":"string","description":"Nombre de usuario de Discord","example":"JohnDoe#1234"},"guildId":{"type":"string","description":"ID del servidor de Discord","example":"987654321098765432"},"birthDate":{"format":"date-time","type":"string","description":"Fecha de nacimiento","example":"2000-01-01T00:00:00.000Z"},"notificationEnabled":{"type":"boolean","description":"Indica si las notificaciones están habilitadas","example":true,"default":false}},"required":["userId","username","guildId","birthDate","notificationEnabled"]},"UpdateBirthdayDto":{"type":"object","properties":{"userId":{"type":"string","description":"ID del usuario de Discord","example":"123456789012345678"},"username":{"type":"string","description":"Nombre de usuario de Discord","example":"JohnDoe#1234"},"guildId":{"type":"string","description":"ID del servidor de Discord","example":"987654321098765432"},"birthDate":{"format":"date-time","type":"string","description":"Fecha de nacimiento","example":"2000-01-01T00:00:00.000Z"},"notificationEnabled":{"type":"boolean","description":"Indica si las notificaciones están habilitadas","example":true,"default":false}}},"ContextInfoDto":{"type":"object","properties":{"channelName":{"type":"string","description":"Nombre del canal","example":"general"},"channelId":{"type":"string","description":"ID del canal","example":"123456789012345678"},"userName":{"type":"string","description":"Nombre del usuario","example":"JohnDoe"},"userId":{"type":"string","description":"ID del usuario","example":"123456789012345678"}}},"GenerateRequestDto":{"type":"object","properties":{"prompt":{"type":"string","description":"Prompt para la IA","example":"¿Cuál es la capital de Francia?"},"guildId":{"type":"string","description":"ID del servidor de Discord","example":"987654321098765432"},"contextInfo":{"description":"Información contextual","allOf":[{"$ref":"#/components/schemas/ContextInfoDto"}]}},"required":["prompt"]},"GuildAIConfig":{"type":"object","properties":{"guildId":{"type":"string","description":"ID del servidor de Discord","example":"987654321098765432"},"systemPrompt":{"type":"string","description":"Instrucciones personalizadas para la IA","example":"Actúa como un asistente amigable y responde en español","default":""},"includeChannelContext":{"type":"boolean","description":"Incluir información del canal en las solicitudes a la IA","example":false,"default":false},"includeUserContext":{"type":"boolean","description":"Incluir información del usuario en las solicitudes a la IA","example":false,"default":false},"modelName":{"type":"string","description":"Nombre del modelo de OpenAI a utilizar","example":"gpt-4o","default":"gpt-4o"},"temperature":{"type":"number","description":"Temperatura para controlar la creatividad de las respuestas","example":0.7,"default":0.7}},"required":["guildId","systemPrompt","includeChannelContext","includeUserContext","modelName","temperature"]},"UpdateAIConfigDto":{"type":"object","properties":{"systemPrompt":{"type":"string","description":"Instrucciones personalizadas para la IA (ej. 'Actúa como un pirata')","example":"Actúa como un asistente amigable y responde en español"},"includeChannelContext":{"type":"boolean","description":"Incluir información del canal en las solicitudes a la IA","example":true},"includeUserContext":{"type":"boolean","description":"Incluir información del usuario en las solicitudes a la IA","example":true},"modelName":{"type":"string","description":"Nombre del modelo de OpenAI a utilizar","example":"gpt-4o"},"temperature":{"type":"number","description":"Temperatura para controlar la creatividad de las respuestas (0-1)","example":0.7,"minimum":0,"maximum":1}}}}}}